+++ apis/r/R/SOMANDArrayBase.R
@@ -19,37 +19,42 @@ SOMANDArrayBase <- R6::R6Class(
   public = list(
 
     #' @description Create a SOMA NDArray named with the URI
-    #' (lifecycle: maturing).
+    #' (lifecycle: maturing).\cr
+    #' \cr
+    #' \strong{Note}: \code{$create()} is considered internal and should not be
+    #' called directly; use factory functions
+    #' (eg. \code{\link{SOMASparseNDArrayCreate}()}) instead.
     #'
     #' @param type An \link[arrow:data-type]{Arrow type} defining the type
     #' of each element in the array.
     #' @param shape a vector of integers defining the shape of the array.
     #' @template param-platform-config
-    #' @param internal_use_only Character value to signal this is a 'permitted'
-    #' call, as \code{create()} is considered internal and should not be called
-    #' directly.
     #'
-    #' @return Returns \code{self}
+    #' @return Returns \code{self}.
     #'
     create = function(
       type,
       shape,
-      platform_config = NULL,
-      internal_use_only = NULL
+      platform_config = NULL
     ) {
-      if (is.null(internal_use_only) || internal_use_only != "allowed_use") {
-        stop(paste(
-          "Use of the create() method is for internal use only. Consider using a",
-          "factory method as e.g. 'SOMASparseNDArrayCreate()'."
-        ), call. = FALSE)
+      envs <- unique(vapply(
+        X = unique(sys.parents()),
+        FUN = function(n) environmentName(environment(sys.function(n))),
+        FUN.VALUE = character(1L)
+      ))
+      if (!"tiledbsoma" %in% envs) {
+        stop(
+          paste(strwrap(private$.internal_use_only("create", "collection")), collapse = '\n'),
+          call. = FALSE
+        )
       }
 
       ## .is_sparse field is being set by dense and sparse private initialisers, respectively
       private$.type <- type # Arrow schema type of data
 
-      dom_ext_tbl <- get_domain_and_extent_array(shape, private$.is_sparse)
+      dom_ext_tbl <- get_domain_and_extent_array(shape, self$is_sparse())
 
-      # Parse the tiledb/create/ subkeys of the platform_config into a handy,
+      # Parse the tiledb/create subkeys of the platform_config into a handy,
       # typed, queryable data structure.
       tiledb_create_options <- TileDBCreateOptions$new(platform_config)
 
@@ -69,22 +74,29 @@ SOMANDArrayBase <- R6::R6Class(
       schema$export_to_c(nasp)
 
       ## create array
-      ctxptr <- super$tiledbsoma_ctx$context()
+      # ctxptr <- self$tiledbsoma_ctx$context()
+      sparse <- if (inherits(self, "SOMASparseNDArray")) {
+        TRUE
+      } else if (inherits(self, "SOMADenseNDArray")) {
+        FALSE
+      } else {
+        stop("Unknown SOMA array type: ", self$class(), call. = FALSE)
+      }
       createSchemaFromArrow(
         uri = self$uri,
         nasp = nasp,
         nadimap = dnaap,
         nadimsp = dnasp,
-        sparse = private$.is_sparse,
-        datatype = if (private$.is_sparse) "SOMASparseNDArray" else "SOMADenseNDArray",
+        sparse = sparse,
+        datatype = if (sparse) "SOMASparseNDArray" else "SOMADenseNDArray",
         pclst = tiledb_create_options$to_list(FALSE),
         ctxxp = private$.soma_context,
         tsvec = self$.tiledb_timestamp_range
       )
       # private$write_object_type_metadata(timestamps)  ## FIXME: temp. commented out -- can this be removed overall?
 
-      self$open("WRITE", internal_use_only = "allowed_use")
-      self
+      self$reopen("WRITE", tiledb_timestamp = self$tiledb_timestamp)
+      return(self)
     },
 
     ## needed eg after open() to set (Arrow) type
@@ -118,7 +130,7 @@ SOMANDArrayBase <- R6::R6Class(
     #'
     #' @return If \code{check_only}, returns the empty string if no error is
     #' detected, else a description of the error. Otherwise, invisibly returns
-    #' \code{NULL}
+    #' \code{NULL}.
     #'
     resize = function(new_shape, check_only = FALSE) {
       stopifnot(
@@ -128,7 +140,13 @@ SOMANDArrayBase <- R6::R6Class(
       )
       # Checking slotwise new shape >= old shape, and <= max_shape, is already done in libtiledbsoma
 
-      reason_string <- resize(self$uri, new_shape, .name_of_function(), check_only, private$.soma_context)
+      reason_string <- resize(
+        uri = self$uri,
+        new_shape = new_shape,
+        function_name_for_messages = .name_of_function(),
+        check_only = check_only,
+        ctxxp = private$.soma_context
+      )
 
       if (isTRUE(check_only)) {
         return(reason_string)
@@ -152,7 +170,7 @@ SOMANDArrayBase <- R6::R6Class(
     #'
     #' @return If \code{check_only}, returns the empty string if no error is
     #' detected, else a description of the error. Otherwise, invisibly returns
-    #' \code{NULL}
+    #' \code{NULL}.
     #'
     tiledbsoma_upgrade_shape = function(shape, check_only = FALSE) {
       stopifnot(
@@ -162,7 +180,13 @@ SOMANDArrayBase <- R6::R6Class(
       )
       # Checking slotwise new shape >= old shape, and <= max_shape, is already done in libtiledbsoma
 
-      reason_string <- tiledbsoma_upgrade_shape(self$uri, shape, .name_of_function(), check_only, private$.soma_context)
+      reason_string <- tiledbsoma_upgrade_shape(
+        uri = self$uri,
+        new_shape = shape,
+        function_name_for_messages = .name_of_function(),
+        check_only = check_only,
+        ctxxp = private$.soma_context
+      )
       if (isTRUE(check_only)) {
         return(reason_string)
       }
@@ -173,7 +197,6 @@ SOMANDArrayBase <- R6::R6Class(
     }
   ),
   private = list(
-    .is_sparse = NULL,
     .type = NULL,
 
     # Given a user-specified shape along a particular dimension, returns a named
@@ -247,9 +270,6 @@ SOMANDArrayBase <- R6::R6Class(
 
       # convert integer to integer64 to match dimension type
       return(bit64::as.integer64(new_shape))
-    },
-
-    # Internal marking of one or zero based matrices for iterated reads
-    zero_based = NA
+    }
   )
 )