+++ libtiledbsoma/src/soma/soma_dense_ndarray.cc
@@ -30,7 +30,7 @@ void SOMADenseNDArray::create(
     auto& index_column_schema = index_columns.second;
     uint64_t index_column_size = index_column_schema->n_children;
 
-    auto schema = std::make_unique<ArrowSchema>();
+    auto schema = make_managed_unique<ArrowSchema>();
     schema->name = nullptr;
     schema->format = strdup("+s");
     schema->n_children = index_column_size + 1;
@@ -80,8 +80,6 @@ void SOMADenseNDArray::create(
 
     SOMAArray::create(
         ctx, uri, tiledb_schema, "SOMADenseNDArray", std::nullopt, timestamp);
-
-    schema->release(schema.get());
 }
 
 std::unique_ptr<SOMADenseNDArray> SOMADenseNDArray::open(
@@ -118,7 +116,7 @@ std::string_view SOMADenseNDArray::soma_data_type() {
 //= public non-static
 //===================================================================
 
-std::unique_ptr<ArrowSchema> SOMADenseNDArray::schema() const {
+managed_unique_ptr<ArrowSchema> SOMADenseNDArray::schema() const {
     return this->arrow_schema();
 }