+++ libtiledbsoma/src/soma/soma_sparse_ndarray.cc
@@ -31,7 +31,7 @@ void SOMASparseNDArray::create(
     auto& index_column_schema = index_columns.second;
     uint64_t index_column_size = index_column_schema->n_children;
 
-    auto schema = std::make_unique<ArrowSchema>();
+    auto schema = make_managed_unique<ArrowSchema>();
     schema->name = nullptr;
     schema->format = strdup("+s");
     schema->n_children = index_column_size + 1;
@@ -81,8 +81,6 @@ void SOMASparseNDArray::create(
 
     SOMAArray::create(
         ctx, uri, tiledb_schema, "SOMASparseNDArray", std::nullopt, timestamp);
-
-    schema->release(schema.get());
 }
 
 std::unique_ptr<SOMASparseNDArray> SOMASparseNDArray::open(
@@ -119,7 +117,7 @@ std::string_view SOMASparseNDArray::soma_data_type() {
 //= public non-static
 //===================================================================
 
-std::unique_ptr<ArrowSchema> SOMASparseNDArray::schema() const {
+managed_unique_ptr<ArrowSchema> SOMASparseNDArray::schema() const {
     return this->arrow_schema();
 }
 }  // namespace tiledbsoma